All miscellaneous files removed
