--- ItsNous-Endgrate-Saaketh-Kesireddy/.browserslistrc
@@ -1,6 +1,0 @@
-Chrome >=79
-ChromeAndroid >=79
-Firefox >=70
-Edge >=79
-Safari >=14
-iOS >=14
