--- ItsNous-Endgrate-Saaketh-Kesireddy/.firebaserc
@@ -1,5 +1,0 @@
-{
-  "projects": {
-    "default": "its-nous-v2"
-  }
-}
