--- ItsNous-Endgrate-Saaketh-Kesireddy/.vscode/extensions.json
@@ -1,5 +1,0 @@
-{
-    "recommendations": [
-        "ionic.ionic"
-    ]
-}
