--- ItsNous-Endgrate-Saaketh-Kesireddy/cypress/e2e/test.cy.ts
@@ -1,6 +1,0 @@
-describe('My First Test', () => {
-  it('Visits the app root url', () => {
-    cy.visit('/')
-    cy.contains('ion-content', 'Tab 1 page')
-  })
-})
