--- ItsNous-Endgrate-Saaketh-Kesireddy/src/App.test.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import App from './App';
-
-test('renders without crashing', () => {
-  const { baseElement } = render(<App />);
-  expect(baseElement).toBeDefined();
-});
