+++ ItsNous-Endgrate-Saaketh-Kesireddy/src/components/SpecificComponents/Finance/News/FinanceNewsCache.tsx
@@ -7,7 +7,7 @@ import {
 
 import { 
   Article 
-} from "../../../@types/article";
+} from "../../../../@types/article";
 
 interface TempArticleViewCache {
   articleGeneralCache: Article[];
