+++ ItsNous-Endgrate-Saaketh-Kesireddy/src/components/SpecificComponents/Finance/News/HighlightedArticle.tsx
@@ -1,4 +1,4 @@
-import { Article } from "../../../@types/article";
+import { Article } from "../../../../@types/article";
 import "../Finance/HighlightedArticle.css";
 
 import {
