+++ ItsNous-Endgrate-Saaketh-Kesireddy/src/pages/Authenticated/Finance/FinanceHome.tsx
@@ -1,0 +1,24 @@
+import { 
+  IonContent,
+  IonPage, 
+} from '@ionic/react';
+
+import BottomNav from '../../../components/GeneralComponents/BottomNav';
+
+import { 
+  FinanceNewsCacheProvider 
+} from '../../../components/SpecificComponents/Finance/News/FinanceNewsCache';
+
+import TotalArticleContainer from '../../../components/SpecificComponents/Finance/News/TotalArticleContainer';
+
+import FinanceHeader from '../../../components/SpecificComponents/Finance/News/FinanceHeader';
+
+const FinanceHome: React.FC = () => {
+  return (
+    <IonPage>
+      test
+    </IonPage>
+  );
+};
+
+export default FinanceHome;
